Fixed: The port connection test may be skipped on some 32-bit platforms

+++ src/validate.c
@@ -1359,7 +1359,7 @@ State_Type check_process(Service_T s) {
                         rv = State_Failed;
                 }
         }
-        int64_t uptimeMilli = s->inf.process->uptime * 1000;
+        int64_t uptimeMilli = (int64_t)(s->inf.process->uptime) * 1000LL;
         for (Port_T pp = s->portlist; pp; pp = pp->next) {
                 //FIXME: instead of pause, try to test, but ignore any errors in the start timeout timeframe ... will allow to display the port response time as soon as available, instead of waiting for 30+ seconds
                 /* pause port tests in the start timeout timeframe while the process is starting (it may take some time to the process before it starts accepting connections) */
